[Refactor] 지저분한 dto 파일 inner class로 정리#49
