+++ contents/todoListAPI/ja/TodoProject/src/main/resources/application-dev.yml
@@ -20,7 +20,7 @@ spring:
       enabled: always
   jpa:
     database-platform: org.hibernate.dialect.MySQL8Dialect
-    open-in-view: true
+    open-in-view: false
     show-sql: true
     hibernate:
       ddl-auto: update
